Add files via upload
This is an Animated Profile Card with Hover Animation using HTML CSS & JavaScript.
At first, on the webpage, there is only an image with round border-color. But when you hover on that card, then the height and width of the card smoothly expand and show the contents of the card.